bot/requirements.txt
aiogram==3.15.0
psycopg2==2.9.10
python_dotenv==1.0.1